pyproject.toml
[build-system]
requires = ["setuptools>=61.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "dirigo_thorlabs_detectors"
version = "0.0.1"
description = "Plugin to use Thorlabs detectors (multiple versions) with Dirigo."
authors = [
    {name = "T. D. Weber", email = "tweber@mit.edu"}
]
license = {text = "MIT"}
dependencies = [
    "dirigo",
    "pyserial",
    "pyvisa"
]

[project.entry-points."dirigo_detectors"]
pda40 = "dirigo_thorlabs_detectors:PDA40"
pmt2100 = "dirigo_thorlabs_detectors:PMT2100"
